Fix #7. Overhauled approach to retrieving arguments. Now traverse the call stack to grab parent function args. Implement 3 new functions for interface, getDotArgs, getNamedArgs, getAllArgs. Each for grabbing different subsets of arguments as described in #7.
To keep this naming scheme consistent, and to indicate that arguments in addition to dots can be converted to flags, dotsToArg has been renamed argsToFlags. Terms used to describe how this package works referrs to "args" as function arguments, and "flags" as commandline flags. I have tried to make these changes to keep things consistent in the documentation with this version as well, although I have assureadly missed a few.

Finally, I added unit tests for the new functions and updated old ones to reflect changes to function names.

N.B. I also redid the internals of argsToFlags which hopefully makes it more clear. Internal list parsing functions in utils.R are useful throughout for dealing with lists anyway.